Create duplicate_test
I was testing out a code that will tell us what files are duplicates and which ones are not. "a" and "b" represents different "folders" and the numbers represent the "files". I was just testing this out before adding a root directory for each folder; just trying to get familiar with the coding concept. I am still not finished adding more codings.
Please feel free to edit and/or add :)